+++ src/containers/BooksList.js
@@ -1,6 +1,6 @@
-import React from 'react-dom';
+import React from 'react';
 import { useSelector } from 'react-redux';
-import Book from './Book';
+import Book from '../components/Book';
 
 function createTableHeader() {
   return (
